Move single use functions into a private file
