+++ draft-packages/popover/KaizenDraft/Popover/types.ts
@@ -5,8 +5,4 @@ export type Variant =
   | "negative"
   | "cautionary"
 
-export type Side = "top" | "bottom" // | "left" | "right" - not yet implemented
-
-export type Position = "start" | "center" | "end"
-
 export type Size = "small" | "large"
